getting global transform moved out of main rendering function
use of DOMMatrix - unstable
considering implementing matrix type from scratch for the engine
